+++ kitty/boss.py
@@ -85,8 +85,6 @@ from .fast_data_types import (
     get_boss,
     get_options,
     get_os_window_size,
-    get_os_window_pos,
-    glfw_primary_monitor_size,
     glfw_get_monitor_workarea,
     global_font_size,
     is_layer_shell_supported,
@@ -1845,9 +1843,14 @@ class Boss:
         else:
             self.mark_os_window_for_close(os_window_id, NO_CLOSE_REQUESTED)
 
-    def on_os_window_closed(self, os_window_id: int, viewport_width: int, viewport_height: int) -> None:
-        self.cached_values['window-size'] = viewport_width, viewport_height
+    def on_os_window_closed(self, os_window_id: int, x: int, y: int, viewport_width: int, viewport_height: int, is_layer_shell: bool) -> None:
         tm = self.os_window_map.pop(os_window_id, None)
+        opts = get_options()
+        if not is_layer_shell:
+            if opts.remember_window_position and not is_wayland() and not self.os_window_map:
+                self.cached_values['window-pos'] = x, y
+                self.cached_values['monitor-workarea'] = glfw_get_monitor_workarea()
+            self.cached_values['window-size'] = viewport_width, viewport_height
         if tm is not None:
             tm.destroy()
         for window_id in tuple(w.id for w in self.window_id_map.values() if getattr(w, 'os_window_id', None) == os_window_id):
@@ -1867,10 +1870,6 @@ class Boss:
             for qt in q:
                 windows += list(qt)
         active_window = self.active_window
-        x, y = get_os_window_pos(active_window.id)
-        monitorGeometryX, monitorGeometryY = glfw_primary_monitor_size()
-        self.cached_values['window-pos'] = x, y
-        self.cached_values['monitor-workarea'] = glfw_get_monitor_workarea()
         msg, num_active_windows = self.close_windows_with_confirmation_msg(windows, active_window)
         x = get_options().confirm_os_window_close[0]
         num = num_active_windows if x < 0 else len(windows)